Prestige Ocean Towers At Marine Lines
Prestige Ocean Towers is a great residential project which is coming up in one of the prime locations of Mumbai. This property is spread over acres of the land region. The housing apartments come in the setup of 2, 3, and 4 BHK. There are perfectly planned skyscraper towers inside the limits of this venture. Prestige Ocean Towers Marine Lines likewise has an arrangement for visitor rooms with a perusing and a television relax. The people group conveniences incorporate a clubhouse, contemplation focus, indoor and outside sports facilities, pool, power, and water back up of 24 hours, vehicle parking, open space, finished garden, firefighting hardware, video security, holding up relax, channeled gas, day care focus, ATM, lift, 3 level security, odds and ends shops and tennis court. The development quality is top-notch and the security is exceptionally guaranteed and is in capable hands. The clinical facilities are not far off and there is imminent availability of this spot with the adjoining regions. The housing format of this spot is open with sufficient regular light and air coursing through. There is similarly a lot of cross ventilation to promise that the spot is windy and new constantly. The environmental factors have a great deal of greenery to add new newness to the existence of occupants remaining here.

Location 

The locality of this spot is Marine Lines, Mumbai. Prestige Ocean Towers Marine Lines, Mumbai has generally essential utilities in close areas. The best of the emergency clinics, parks, schools, establishments, shopping centers, multiplexes, and so on are not far off. There is more than adequate access to public vehicles and there is a ton of simplicity in driving to all kinds of spots from here.

For More Info:
Visit: https://www.prlog.org/12939391-quality-living-starts-here-at-prestige-ocean-towers-mumbai.html